Cleanup ARM-gcc compiler warnings.

+++ src/NimBLERemoteCharacteristic.cpp
@@ -563,11 +563,11 @@ int NimBLERemoteCharacteristic::onReadCB(uint16_t conn_handle,
 
     if(rc == 0) {
         if(attr) {
-            uint32_t data_len = OS_MBUF_PKTLEN(attr->om);
+            uint16_t data_len = OS_MBUF_PKTLEN(attr->om);
             if(((*strBuf).length() + data_len) > BLE_ATT_ATTR_MAX_LEN) {
                 rc = BLE_ATT_ERR_INVALID_ATTR_VALUE_LEN;
             } else {
-                NIMBLE_LOGD(LOG_TAG, "Got %d bytes", data_len);
+                NIMBLE_LOGD(LOG_TAG, "Got %u bytes", data_len);
                 (*strBuf) += std::string((char*) attr->om->om_data, data_len);
                 return 0;
             }
